build up click functionality
