+++ client/src/App.js
@@ -1,10 +1,12 @@
 import React, { Component } from "react";
 import "./App.css";
+import Ideas from './pages/Ideas';
 
 class App extends Component {
   render() {
     return (
       <div className="App">
+        <Ideas />
       </div>
     );
   }
