+++ client/src/components/CampaignDisplay/index.js
@@ -2,16 +2,13 @@ import React from 'react';
 // import API from '../utils/API';
 
 const CampaignDisplay = (props) => {
-  const { title, author, synopsis } = props;
+  const { title, author, synopsis, onClick, campaignID } = props;
   return(
-    <div>
-      <header>New Campaigns</header>
-      <section id="campaignDisplay">
-        <h2>Title: {title}</h2>
-        <h3>Author: {author}</h3>
-        <p>{synopsis}</p>
-      </section>
-    </div>
+    <section id="campaignDisplay" onClick={onClick} data-id={campaignID}>
+      <h2>Title: {title}</h2>
+      <h3>Author: {author}</h3>
+      <p>{synopsis}</p>
+    </section>
   );
 };
