+++ client/src/utils/API.js
@@ -1,16 +1,32 @@
 import axios from 'axios';
 
 export default {
+  // Post a campaign
   campaignPost: (campaign) => {
     return axios.post('/api/campaign', campaign);
   },
+  // Get all campaigns
   campaignGet: () => {
     return axios.get('/api/campaign');
   },
-  discussionPost: (discussion) => {
-    return axios.post('/api/discussion', discussion);
+  // Gets all votes
+  getVotes: function () {
+    return axios.get("/api/vote");
   },
-  discussionGet: () => {
-    return axios.get('/api/discussion');
+  // Gets the vote with the given id
+  getvote: function (id) {
+      return axios.get("/api/vote/" + id);
+  },
+  // Deletes the vote with the given id
+  deleteVote: function (id) {
+      return axios.delete("/api/vote/" + id);
+  },
+  // Saves a vote to the database
+  saveVote: function (voteData) {
+      return axios.post("/api/vote", voteData);
+  },
+  // updates a vote
+  updateVote: function (id, voteData) {
+      return axios.put("/api/vote/" + id, voteData);
   }
 };
