+++ routes/api/index.js
@@ -5,9 +5,6 @@ const voteRoutes = require("./vote");
 // Campaign routes
 router.use("/campaign", campaignRoutes);
 
-// Discussion routes
-router.use("/discussion", discussionRoutes);
-
 // vote routes
 router.use("/vote", voteRoutes);
